refactor : OutputView 클래스 분리
